+++ src/app.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common'
 
 import { configModule } from '~/core/config.module'
+import { databaseModule } from '~/core/database/database.module'
 import { OtherModule } from '~/other/other.module'
 
 @Module({
-  imports: [configModule, OtherModule],
+  imports: [configModule, databaseModule, OtherModule],
 })
 export class AppModule {}