+++ src/core/config.module.ts
@@ -16,6 +16,14 @@ export const config = () => ({
   http: {
     port: Number(env.HTTP_PORT),
   },
+
+  db: {
+    host: env.DB_HOST,
+    port: Number(env.DB_PORT),
+    database: env.DB_NAME,
+    user: env.DB_USER,
+    password: env.DB_PASS,
+  },
 })
 
 export const configModule = ConfigModule.forRoot({
@@ -27,5 +35,11 @@ export const configModule = ConfigModule.forRoot({
       .default('development'),
 
     HTTP_PORT: Joi.number().default('3000'),
+
+    DB_HOST: Joi.string().empty('').default('localhost'),
+    DB_PORT: Joi.number().empty('').default('5432'),
+    DB_NAME: Joi.string().required(),
+    DB_USER: Joi.string().required(),
+    DB_PASS: Joi.string().required(),
   }),
 })