+++ src/core/database/database.service.ts
@@ -1,0 +1,5 @@
+import { NodePgDatabase } from 'drizzle-orm/node-postgres'
+
+import type * as schema from '~/core/database/schema'
+
+export class DatabaseService extends NodePgDatabase<typeof schema> {}