+++ src/core/database/index.ts
@@ -1,0 +1,3 @@
+export { DatabaseService } from './database.service'
+
+export * from './schema'